Select dynamically generated content
Since the food names and calories are not yet on the page when it's
loaded (since they are dynamically generated by a user), they will not
get this event listener attached.  In order to attach the event listener
so that pressing the 'Enter' key removes the focus from the editable
area, we have to select the parent instead.