Persist edited food names and calories
When a user clicks on a food name or calorie entry
and edits that field, upon pressing ‘Enter’, the 
change will persist - even after page refreshes.

Partial functionality only:
This works for hitting the ‘Enter’ key but it
doesn’t work if the user simply clicks away from
the input field.